[DM-15571] Turn kube/prod into the prod deploy.
We need to change the namespace to dax-prod, and also provide a
volume with a different name that contains prod in it.

This is because the volumes are not namespaced, but the claims are,
and the volumes can't be shared across namespaces.

I could have dev and prod instances in the same namespace, but this
implies they might be using different claims and names all over the
place, and can introduce unintended consequences.  I'd like to
keep it to different namespaces that are isolated from each other
if possible.