[DM-15571] Configure dev environment.
1) Change namespace over from dax-prod to dax-dev.
2) Create a volume with a different name, since this is the only
non-namespaced k8s object, and they can't be shared across namespaces.
3) Fix the prod/destroy.sh script to destroy the correctly named pv
for prod, and delete the appropriate dev one for the destroy.sh in dev.
4) Modify the ingress rule to route everything from /dev for now, and
rewrite the URL.  I haven't found a way to configure the path for webserv
yet, but maybe this can be improved.  The prod version doesn't need this
because we are serving it at its natural URL.