[DM-15396] Stack container: checkpoint.
Got everything mostly stitched together.  Things run as the lsst
user, and we're using the stack container.  Install python modules
on their branches where I've been working from.

One last problem is getting the imgserv tests to run as part of
creating the container, and then after the merge the branches
can be changed to master.